Working page title translation!
Also, moved mlang translation function to AppRuntime and AppExtension, removed TranslationExtension and TranslationRuntime files
