+++ src/Twig/DateExtension.php
@@ -15,7 +15,6 @@ class DateExtension extends AbstractExtension
         return [
             new TwigFilter('timeAgo', [DateRuntime::class, 'timeAgo']),
             new TwigFilter('timeDiff', [DateRuntime::class, 'timeDiff']),
-            new TwigFilter('transMlang', [DateRuntime::class, 'transMlang']),
         ];
     }
