--- src/Twig/TranslationExtension.php
@@ -1,18 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Twig;
-
-use Twig\Extension\AbstractExtension;
-use Twig\TwigFilter;
-
-class TranslationExtention extends AbstractExtension
-{
-    public function getFilters(): array
-    {
-        return [
-            new TwigFilter('transMlang', [TranslationRuntime::class, 'transMlang'])
-        ];
-    }
-}
